Suite de la suppression du TypeScript dans les fichiers Js liés à Vue
