+++ client/src/components/MaterialsListEditor/ReuseEventMaterials/index.js
@@ -1,7 +1,7 @@
 import './index.scss';
 import { ref, computed } from '@vue/composition-api';
-import useI18n from '@/hooks/useI18n';
-import useRouter from '@/hooks/useRouter';
+import useI18n from '@/hooks/vue/useI18n';
+import useRouter from '@/hooks/vue/useRouter';
 import Loading from '@/components/Loading';
 import ErrorMessage from '@/components/ErrorMessage/index';
 import MaterialsSorted from '@/components/MaterialsSorted';
