+++ client/src/hooks/vue/useI18n.js
@@ -1,8 +1,6 @@
 import getRuntimeVm from '@/utils/getRuntimeVm';
 
-import type { I18nTranslate } from 'vuex-i18n';
-
-const useI18n = (): I18nTranslate => {
+const useI18n = () => {
     const vm = getRuntimeVm();
     return getRuntimeVm().$t.bind(vm);
 };
