+++ client/src/hooks/vue/useRoutePage.js
@@ -1,17 +1,15 @@
 import { computed } from '@vue/composition-api';
-import useRouter from '@/hooks/useRouter';
+import useRouter from '@/hooks/vue/useRouter';
 
-import type { Ref } from '@vue/composition-api';
-
-const useRoutePage = (): Ref<number> => {
+const useRoutePage = () => {
     const { route } = useRouter();
 
-    const page = computed<number>(() => {
+    const page = computed(() => {
         const { query } = route.value;
         if (!query || !query.page) {
             return 1;
         }
-        return Number.parseInt(query.page as string, 10);
+        return Number.parseInt(query.page, 10);
     });
 
     return page;
