+++ client/src/utils/errors.js
@@ -1,33 +1,28 @@
 import axios from 'axios';
 
-import type { AxiosError } from 'axios';
-import type { I18nTranslate } from 'vuex-i18n';
-import type { FormErrorDetail } from '@/stores/api/@types';
-
-const isApiErrorCode = (error: unknown, code: number): boolean => {
+const isApiErrorCode = (error, code) => {
     if (!axios.isAxiosError(error)) {
         return false;
     }
-
     return (error.response?.status === code);
 };
 
-const getValidationErrors = (error: unknown): FormErrorDetail | null => {
+const getValidationErrors = (error) => {
     if (!isApiErrorCode(error, 400)) {
         return null;
     }
 
-    const { details } = (error as AxiosError).response?.data?.error || { details: {} };
+    const { details } = error.response?.data?.error || { details: {} };
     return details;
 };
 
-const getErrorMessage = (error: unknown, __: I18nTranslate): string => {
+const getErrorMessage = (error, __) => {
     if (typeof error === 'string') {
         return error;
     }
 
     if (!axios.isAxiosError(error)) {
-        const message = (error as Error).message || 'unknown';
+        const message = error.message || 'unknown';
         return __('errors.generic', { message });
     }
